Move implementation from examples_level_zero.h to examples_level_zero.c
