+++ examples/ipc_level_zero/ipc_level_zero.c
@@ -8,6 +8,7 @@
  */
 
 #include <stdio.h>
+#include <stdlib.h>
 
 #include "umf/ipc.h"
 #include "umf/memory_pool.h"
