+++ examples/level_zero_shared_memory/level_zero_shared_memory.c
@@ -7,6 +7,8 @@
  *
  */
 
+#include <stdio.h>
+
 #include <umf/memory_pool.h>
 #include <umf/pools/pool_disjoint.h>
 #include <umf/providers/provider_level_zero.h>
